ALMA_streamlit/requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-openai
langchain-pinecone
pinecone-client
pydub
edge-tts
pygame
speechrecognition
langsmith